chore: refactor labels functions into service
