+++ src/Components/ServiceScene/Breakdowns/LabelBreakdownScene.test.ts
@@ -1,6 +1,6 @@
 import { AdHocFiltersVariable, sceneGraph, SceneObject } from '@grafana/scenes';
 import { VAR_FIELDS, VAR_LABEL_GROUP_BY_EXPR } from '../../../services/variables';
-import { buildLabelsQuery } from './LabelBreakdownScene';
+import { buildLabelsQuery } from '../../../services/labels';
 
 describe('buildLabelsQuery', () => {
   test('should build no-parser query with no filters', () => {
